backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.1.0
django-filter==23.5
drf-yasg==1.21.7
cryptography==41.0.7
PyJWT==2.8.0

# OAuth2 and Email Providers
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
msal==1.25.0
requests==2.31.0

# AI/ML Libraries
google-generativeai>=0.8.0

